Refactor Composite into separate module.
